+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.0.1: 2025-08-30
+
+* Split environment configuration into development and production files following Mastodon conventions
+* Add version headers to scripts
+
 ### 1.0.0: 2025-08-30
 
 * Release 1.0.0